+++ libs/render_delegate/camera.cpp
@@ -38,9 +38,6 @@ HdArnoldCamera::HdArnoldCamera(HdArnoldRenderDelegate* renderDelegate, const Sdf
     // We create a persp_camera by default and optionally replace the node in ::Sync.
     _camera = renderDelegate->CreateArnoldNode(str::persp_camera, AtString(id.GetText()));
     _delegate = renderDelegate;
-    if (!id.IsEmpty()) {
-        AiNodeSetStr(_camera, str::name, AtString(id.GetText()));
-    }
 }
 
 HdArnoldCamera::~HdArnoldCamera() {
@@ -54,7 +51,7 @@ HdArnoldCamera::~HdArnoldCamera() {
         if (_camera == AiNodeGetPtr(options, str::subdiv_dicing_camera))
             AiNodeResetParameter(options, str::subdiv_dicing_camera);        
 
-        AiNodeDestroy(_camera); 
+        _delegate->DestroyArnoldNode(_camera);
     }
 }